Extract HTTP service into dedicated package and UI plugin
