--- DesktopApplicationTemplate.UI/Views/Http/Edit/HttpEditServiceView.xaml.cs
@@ -1,22 +1,0 @@
-using System.Windows.Controls;
-using DesktopApplicationTemplate.Core.Services;
-using DesktopApplicationTemplate.UI.ViewModels.Http.Edit;
-
-namespace DesktopApplicationTemplate.UI.Views.Http.Edit;
-
-public partial class HttpEditServiceView : Page
-{
-    private readonly ILoggingService _logger;
-
-    public HttpEditServiceView(ILoggingService logger)
-    {
-        InitializeComponent();
-        _logger = logger;
-    }
-
-    public void Initialize(HttpEditServiceViewModel vm)
-    {
-        DataContext = vm;
-        vm.Logger = _logger;
-    }
-}
