--- DesktopApplicationTemplate.UI/Views/Http/HttpServiceView.xaml
@@ -1,114 +1,0 @@
-﻿<Page x:Class="DesktopApplicationTemplate.UI.Views.Http.HttpServiceView" 
-      xmlns:vm="clr-namespace:DesktopApplicationTemplate.UI.ViewModels.Http"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:local="clr-namespace:DesktopApplicationTemplate.UI.Views"
-      xmlns:helpers="clr-namespace:DesktopApplicationTemplate.UI.Helpers"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:behaviors="clr-namespace:DesktopApplicationTemplate.UI.Behaviors"
-
-      xmlns:shared="clr-namespace:DesktopApplicationTemplate.UI.Views.Shared"
-
-      mc:Ignorable="d" Background="#f3f3f3">
-
-    <Page.Resources>
-        <helpers:StringNullOrEmptyToVisibilityConverter x:Key="StringNullOrEmptyToVisibilityConverter" />
-        <SolidColorBrush x:Key="PrimaryColor" Color="#90ee90"/>
-        <SolidColorBrush x:Key="SecondaryColor" Color="#d3d3d3"/>
-        <SolidColorBrush x:Key="DangerColor" Color="#ff6961"/>
-        <SolidColorBrush x:Key="WarningColor" Color="#f8d568"/>
-        <helpers:NullToVisibilityConverter x:Key="NullToVisibilityConverter" />
-    </Page.Resources>
-
-
-    <Grid Margin="20">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-        </Grid.RowDefinitions>
-
-        <Border Grid.Row="0" CornerRadius="8" BorderThickness="1" Padding="10" Margin="0,0,0,10"
-                Background="{Binding Tag.BackgroundColor, RelativeSource={RelativeSource AncestorType=Page}}"
-                BorderBrush="{Binding Tag.BorderColor, RelativeSource={RelativeSource AncestorType=Page}}"
-                ToolTip="{Binding Tag.DescriptorTooltip, RelativeSource={RelativeSource AncestorType=Page}}">
-            <DockPanel>
-                <Border Width="40" Height="40" CornerRadius="20" Margin="0,0,10,0"
-                        Background="{Binding Tag.BorderColor, RelativeSource={RelativeSource AncestorType=Page}}"
-                        Visibility="{Binding Tag.IconGlyph, RelativeSource={RelativeSource AncestorType=Page}, Converter={StaticResource NullToVisibilityConverter}}">
-                    <TextBlock Text="{Binding Tag.IconGlyph, RelativeSource={RelativeSource AncestorType=Page}}"
-                               FontSize="22" HorizontalAlignment="Center" VerticalAlignment="Center" />
-                </Border>
-                <StackPanel>
-                    <TextBlock Text="{Binding Tag.DisplayName, RelativeSource={RelativeSource AncestorType=Page}}" FontWeight="Bold" />
-                    <TextBlock Text="{Binding Tag.DescriptorDescription, RelativeSource={RelativeSource AncestorType=Page}}"
-                               Foreground="Gray" TextWrapping="Wrap"
-                               Visibility="{Binding Tag.DescriptorDescription, RelativeSource={RelativeSource AncestorType=Page}, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}" />
-                </StackPanel>
-            </DockPanel>
-        </Border>
-
-        <StackPanel Grid.Row="1" Orientation="Horizontal" Margin="0 0 0 10">
-            <ComboBox Width="100" ItemsSource="{Binding Methods}" SelectedItem="{Binding SelectedMethod}" />
-            <Grid Width="400" Margin="10 0">
-                <TextBox Text="{Binding Url, UpdateSourceTrigger=PropertyChanged, ValidatesOnNotifyDataErrors=True}" Name="UrlTextBox"
-                         behaviors:TextBoxHintBehavior.AutoToolTip="True"/>
-                <TextBlock Text="http://127.0.0.1/api"
-               IsHitTestVisible="False"
-               Foreground="Gray"
-               VerticalAlignment="Center"
-               Margin="5,0,0,0"
-               Visibility="{Binding Text, ElementName=UrlTextBox, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}"/>
-            </Grid>
-            <Button Content="Send" Width="100" Command="{Binding SendCommand}" Background="{StaticResource PrimaryColor}"/>
-            <Button Content="Help" Width="80" Margin="5,0,0,0" Click="Help_Click" Background="{StaticResource SecondaryColor}"/>
-        </StackPanel>
-
-        <DataGrid Grid.Row="2" ItemsSource="{Binding Headers}" AutoGenerateColumns="False" SelectedItem="{Binding SelectedHeader}" Margin="0 0 0 10" Height="100" MaxWidth="300">
-            <DataGrid.Columns>
-                <DataGridTextColumn Header="Key" Binding="{Binding Key}" Width="*"/>
-                <DataGridTextColumn Header="Value" Binding="{Binding Value}" Width="*"/>
-            </DataGrid.Columns>
-        </DataGrid>
-        <StackPanel Orientation="Horizontal" Grid.Row="3" HorizontalAlignment="Right" Margin="0,5,0,0">
-            <Button Content="Add" Width="50" Command="{Binding AddHeaderCommand}" Background="{StaticResource PrimaryColor}"/>
-            <Button Content="Remove" Width="60" Margin="5,0,0,0" Command="{Binding RemoveHeaderCommand}" Background="{StaticResource DangerColor}"/>
-        </StackPanel>
-
-        <Grid Grid.Row="4">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*"/>
-                <ColumnDefinition Width="*"/>
-            </Grid.ColumnDefinitions>
-
-            <Grid Grid.Column="0">
-                <TextBox Text="{Binding RequestBody}" x:Name="RequestBodyBox" AcceptsReturn="True" VerticalScrollBarVisibility="Auto"
-                         behaviors:TextBoxHintBehavior.AutoToolTip="True" />
-                <TextBlock Text="Request Body" IsHitTestVisible="False" Foreground="Gray" Margin="5,0,0,0"
-                           VerticalAlignment="Center"
-                           Visibility="{Binding Text, ElementName=RequestBodyBox, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}"/>
-            </Grid>
-            <StackPanel Grid.Column="1">
-                <Grid>
-                    <TextBox Text="{Binding ResponseBody}" x:Name="ResponseBodyBox" AcceptsReturn="True" VerticalScrollBarVisibility="Auto" IsReadOnly="True" Height="120"
-                             behaviors:TextBoxHintBehavior.AutoToolTip="True"/>
-                    <TextBlock Text="Response Body" IsHitTestVisible="False" Foreground="Gray" Margin="5,0,0,0"
-                               VerticalAlignment="Center"
-                               Visibility="{Binding Text, ElementName=ResponseBodyBox, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}"/>
-                </Grid>
-                <TextBlock Text="Status: {Binding StatusCode}" Margin="0,5,0,0"/>
-            </StackPanel>
-        </Grid>
-
-        <shared:ServiceLogView Grid.Row="5" Margin="0,10,0,0" x:Name="LogView" />
-
-        <StackPanel Grid.Row="6" HorizontalAlignment="Right" Margin="0,10,0,0">
-            <Button Content="Save Configuration" Command="{Binding SaveCommand}" Width="150" Background="{StaticResource PrimaryColor}"/>
-        </StackPanel>
-    </Grid>
-</Page>
