+++ Services/Http/HttpPackage/plugin.manifest.json
@@ -1,0 +1,12 @@
+{
+  "id": "DesktopApplicationTemplate.Http",
+  "name": "HTTP Service Package",
+  "version": "1.0.0",
+  "entryAssembly": "DesktopApplicationTemplate.Http.dll",
+  "serviceAssemblies": [
+    "DesktopApplicationTemplate.Http.dll"
+  ],
+  "probingPaths": [
+    "libs"
+  ]
+}
