+++ Services/Http/HttpUI/Views/Http/Create/HttpCreateServiceView.xaml.cs
@@ -1,0 +1,15 @@
+using System.Windows.Controls;
+using DesktopApplicationTemplate.Core.Services;
+using DesktopApplicationTemplate.Services.Http.UI.ViewModels.Http.Create;
+
+namespace DesktopApplicationTemplate.Services.Http.UI.Views.Http.Create;
+
+public partial class HttpCreateServiceView : Page
+{
+    public HttpCreateServiceView(HttpCreateServiceViewModel vm, ILoggingService logger)
+    {
+        InitializeComponent();
+        DataContext = vm;
+        vm.Logger = logger;
+    }
+}
