+++ Services/Http/HttpUI/Views/Http/Edit/HttpEditServiceView.xaml
@@ -1,0 +1,33 @@
+<Page x:Class="DesktopApplicationTemplate.Services.Http.UI.Views.Http.Edit.HttpEditServiceView"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:views="clr-namespace:DesktopApplicationTemplate.UI.Views"
+      mc:Ignorable="d">
+    <Grid Margin="20">
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+        </Grid.RowDefinitions>
+
+        <TextBlock Grid.Row="0" Grid.Column="0" Text="Service Name" Style="{StaticResource FormLabel}"/>
+        <TextBox Grid.Row="0" Grid.Column="1" Text="{Binding ServiceName}" Style="{StaticResource FormField}"/>
+
+        <TextBlock Grid.Row="1" Grid.Column="0" Text="Base URL" Style="{StaticResource FormLabel}"/>
+        <TextBox Grid.Row="1" Grid.Column="1" Text="{Binding BaseUrl}" Style="{StaticResource FormField}"/>
+
+        <views:EditorButtonBar Grid.Row="2" Grid.ColumnSpan="2"
+                               AdvancedConfigCommand="{Binding AdvancedConfigCommand}"
+                               SaveCommand="{Binding SaveCommand}"
+                               CancelCommand="{Binding CancelCommand}"
+                               AdvancedAutomationName="Open HTTP Advanced Configuration"
+                               SaveAutomationName="Save HTTP Service"
+                               CancelAutomationName="Cancel HTTP Edit" />
+    </Grid>
+</Page>
